+++ main/ParamsChecker.h
@@ -7,7 +7,7 @@ namespace directgraph {
     private:
     public:
         ParamsChecker();
-        bool checkEllipse(float x, float y, float rx, float ry);
+        bool checkEllipse(int32_t x, int32_t y, uint32_t rx, uint32_t ry);
         bool checkPixel(int_fast32_t x, int_fast32_t y, uint_fast32_t maxWidth, uint_fast32_t maxHeight);
         void checkFillStyle(uint_fast8_t fillstyle);
     };
