+++ tests/dx9/dx9_fillellipse_old.cpp
@@ -13,8 +13,8 @@ public:
     NiceMock<QueueReaderStub> _readerStub;
     IMyWindow *win;
     IFeatures::ShaderVersion noShVer;
-    static float WIDTH;
-    static float HEIGHT;
+    static int32_t WIDTH;
+    static int32_t HEIGHT;
 
     DX9FillEllipseOldTest() {
         IFeatures::ShaderVersion t = {0, 0};
@@ -31,15 +31,15 @@ public:
     }
 };
 
-float DX9FillEllipseOldTest::WIDTH = 200;
-float DX9FillEllipseOldTest::HEIGHT = 200;
+int32_t DX9FillEllipseOldTest::WIDTH = 200;
+int32_t DX9FillEllipseOldTest::HEIGHT = 200;
 
 IMG_TEST_F(DX9FillEllipseOldTest, SimpleEllipse){
     QueueItem items[] = {
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -50,7 +50,7 @@ IMG_TEST_F(DX9FillEllipseOldTest, SimpleEllipseReverse){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/4, WIDTH/2, HEIGHT/4)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/4, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/4))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -61,13 +61,13 @@ IMG_TEST_F(DX9FillEllipseOldTest, Ellipse4){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/2, HEIGHT/2),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/2)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/5, WIDTH/4, HEIGHT/5),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/5, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/5)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0xFF0000),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/5, WIDTH/4, HEIGHT/5),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/5, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/5)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0xFFFF00),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, 4*HEIGHT/5, WIDTH/2, HEIGHT/5)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, 4*HEIGHT/5, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/5))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -78,11 +78,11 @@ IMG_TEST_F(DX9FillEllipseOldTest, EllipseThenBar2){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/2, HEIGHT/2),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/2)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0xFF0000),
             QueueItemCreator::create<QueueItem::BAR>(0, 0, WIDTH/8, HEIGHT/8),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/5, WIDTH/4, HEIGHT/5),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/5, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/5)),
             QueueItemCreator::create<QueueItem::BAR>(WIDTH/2, HEIGHT/2, 3*WIDTH/4, 3*HEIGHT/4)
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
@@ -94,7 +94,7 @@ IMG_TEST_F(DX9FillEllipseOldTest, Ellipse1px){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(1, 1, 1, 1)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(1, 1, 1u, 1u)
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -105,15 +105,15 @@ IMG_TEST_F(DX9FillEllipseOldTest, Ellipse2to10px){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(10, 10, 2, 2),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(20, 20, 3, 3),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(30, 30, 4, 4),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(40, 40, 5, 5),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(50, 50, 6, 6),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(65, 65, 7, 7),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(80, 80, 8, 8),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(95, 95, 9, 9),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(110, 110, 10, 10)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(10, 10, 2u, 2u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(20, 20, 3u, 3u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(30, 30, 4u, 4u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(40, 40, 5u, 5u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(50, 50, 6u, 6u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(65, 65, 7u, 7u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(80, 80, 8u, 8u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(95, 95, 9u, 9u),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(110, 110, 10u, 10u)
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -124,7 +124,7 @@ IMG_TEST_F(DX9FillEllipseOldTest, EllipseLarge){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x00BBFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/2, HEIGHT/2)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/2))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
