+++ tests/dx9/dx9_fillellipse_style.cpp
@@ -12,8 +12,8 @@ class DX9FillEllipseStyleTest: public ImageTest, public CommonSimple{
 public:
     NiceMock<QueueReaderStub> _readerStub;
     IMyWindow *win;
-    static float WIDTH;
-    static float HEIGHT;
+    static int32_t WIDTH;
+    static int32_t HEIGHT;
 
     DX9FillEllipseStyleTest() {
         win = createWindow(WIDTH, HEIGHT);
@@ -24,15 +24,15 @@ public:
     }
 };
 
-float DX9FillEllipseStyleTest::WIDTH = 200;
-float DX9FillEllipseStyleTest::HEIGHT = 200;
+int32_t DX9FillEllipseStyleTest::WIDTH = 200;
+int32_t DX9FillEllipseStyleTest::HEIGHT = 200;
 
 IMG_TEST_F(DX9FillEllipseStyleTest, SimpleEllipseHatch){
     QueueItem items[] = {
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(HATCH_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -43,8 +43,8 @@ IMG_TEST_F(DX9FillEllipseStyleTest, TwoEllipsesHatch){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(HATCH_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2)),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -55,13 +55,13 @@ IMG_TEST_F(DX9FillEllipseStyleTest, EllipsesDifferentFill){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(HATCH_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(HATCH_FILL, 0xAA2200),
             QueueItemCreator::create<QueueItem::BGCOLOR>(0x00FF00),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/4, HEIGHT/4),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/4)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SLASH_FILL, 0xFF00FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/4, WIDTH/8, HEIGHT/5)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2)),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/4, static_cast<uint32_t>(WIDTH/8), static_cast<uint32_t>(HEIGHT/5))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
@@ -72,18 +72,18 @@ IMG_TEST_F(DX9FillEllipseStyleTest, EllipsesDifferentFillAndSolid){
             QueueItemCreator::create<QueueItem::CLEAR>(),
             QueueItemCreator::create<QueueItem::SETLINESTYLE>(NULL_LINE, 0, 0),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/2, HEIGHT/2),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/2)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(HATCH_FILL, 0x0000FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(HATCH_FILL, 0xAA2200),
             QueueItemCreator::create<QueueItem::BGCOLOR>(0x00FF00),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/4, HEIGHT/4),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/4)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SLASH_FILL, 0xFF00FF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/4, WIDTH/8, HEIGHT/5),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, WIDTH/4, HEIGHT/2),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2)),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/4, HEIGHT/4, static_cast<uint32_t>(WIDTH/8), static_cast<uint32_t>(HEIGHT/5)),
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(3*WIDTH/4, HEIGHT/2, static_cast<uint32_t>(WIDTH/4), static_cast<uint32_t>(HEIGHT/2)),
             QueueItemCreator::create<QueueItem::SETFILLSTYLE>(SOLID_FILL, 0xFFAAFF),
-            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, WIDTH/2, HEIGHT/5)
+            QueueItemCreator::create<QueueItem::FILLELLIPSE>(WIDTH/2, HEIGHT/2, static_cast<uint32_t>(WIDTH/2), static_cast<uint32_t>(HEIGHT/5))
     };
     _readerStub.addItems(items, sizeof(items) / sizeof(QueueItem));
     return afterTestSimple(win, &_readerStub);
